First Version of the Gripper Optimization Functions Using MATLAB
These five functions are the first version of the gripper optimization where we use MATLAB to plot how the Reaction Forces at the Toes of the gripper changes as a function of some varying Link Length. It is important to note that at this time, the code only attempts to observe how the reaction for is influenced by a either one varying link while all other links are assumed to have a constant length or two varying links (a link symmetric pair where the link pair both change to maintain structural symmetry in the link design).